Relocated some files for better organisation
