+++ lib/screens/home/event_list/UncompletedPastEventList.dart
@@ -1,21 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:plannus/models/Event.dart';
-import 'package:plannus/screens/home/EventTile.dart';
-import 'package:plannus/screens/home/EmptyCardWithText.dart';
+import 'package:plannus/screens/home/home_elements/EventTile.dart';
+import 'package:plannus/screens/home/home_elements/EmptyCardWithText.dart';
 
-class UnfinishedEventList extends StatefulWidget {
-  const UnfinishedEventList({Key key}) : super(key: key);
+class UncompletedPastEventList extends StatefulWidget {
+  const UncompletedPastEventList({Key key}) : super(key: key);
 
   @override
-  _UnfinishedEventListState createState() => _UnfinishedEventListState();
+  _UncompletedPastEventList createState() => _UncompletedPastEventList();
 }
 
-class _UnfinishedEventListState extends State<UnfinishedEventList> {
+class _UncompletedPastEventList extends State<UncompletedPastEventList> {
   @override
   Widget build(BuildContext context) {
-
-    int comparator(Event a, Event b) {
+    int comparator(Event b, Event a) {
       int temp = a.startTime.compareTo(b.startTime);
       return temp != 0
           ? temp
@@ -23,9 +22,9 @@ class _UnfinishedEventListState extends State<UnfinishedEventList> {
     }
 
     List<Event> events = Provider.of<List<Event>>(context) ?? [];
-    events = events.where((Event e) => e.endTime.compareTo(DateTime.now()) > 0 && !e.passed).toList();
+    events = events.where((Event e) => e.passed && !e.completed).toList();
     if (events.length == 0) {
-      return EmptyCardWithText(text: "No upcoming activities. What are you waiting for?");
+      return EmptyCardWithText(text: "No activities to show.");
     }
     events.sort(comparator);
     events = events.length > 10 ? events.sublist(0, 10) : events;
@@ -34,7 +33,7 @@ class _UnfinishedEventListState extends State<UnfinishedEventList> {
       scrollDirection: Axis.horizontal,
       itemCount: events.length,
       itemBuilder: (context, index) {
-        return EventTile(event: events[index]);
+        return EventTile(event: events[index],);
       },
     );
   }
