+++ lib/screens/home/event_list/UnfinishedEventList.dart
@@ -1,20 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:plannus/models/Event.dart';
-import 'package:plannus/screens/home/EventTile.dart';
-import 'package:plannus/screens/home/EmptyCardWithText.dart';
+import 'package:plannus/screens/home/home_elements/EventTile.dart';
+import 'package:plannus/screens/home/home_elements/EmptyCardWithText.dart';
 
-class MaybeFinishedEventList extends StatefulWidget {
-  const MaybeFinishedEventList({Key key}) : super(key: key);
+class UnfinishedEventList extends StatefulWidget {
+  const UnfinishedEventList({Key key}) : super(key: key);
 
   @override
-  _MaybeFinishedEventListState createState() => _MaybeFinishedEventListState();
+  _UnfinishedEventListState createState() => _UnfinishedEventListState();
 }
 
-class _MaybeFinishedEventListState extends State<MaybeFinishedEventList> {
+class _UnfinishedEventListState extends State<UnfinishedEventList> {
   @override
   Widget build(BuildContext context) {
-    int comparator(Event b, Event a) {
+
+    int comparator(Event a, Event b) {
       int temp = a.startTime.compareTo(b.startTime);
       return temp != 0
           ? temp
@@ -22,9 +23,9 @@ class _MaybeFinishedEventListState extends State<MaybeFinishedEventList> {
     }
 
     List<Event> events = Provider.of<List<Event>>(context) ?? [];
-    events = events.where((Event e) => e.endTime.compareTo(DateTime.now()) <= 0 && !e.passed).toList();
+    events = events.where((Event e) => e.endTime.compareTo(DateTime.now()) > 0 && !e.passed).toList();
     if (events.length == 0) {
-      return EmptyCardWithText(text: "No activities to show.");
+      return EmptyCardWithText(text: "No upcoming activities. What are you waiting for?");
     }
     events.sort(comparator);
     events = events.length > 10 ? events.sublist(0, 10) : events;
@@ -33,7 +34,7 @@ class _MaybeFinishedEventListState extends State<MaybeFinishedEventList> {
       scrollDirection: Axis.horizontal,
       itemCount: events.length,
       itemBuilder: (context, index) {
-        return EventTile(event: events[index],);
+        return EventTile(event: events[index]);
       },
     );
   }
