+++ lib/screens/home/home_elements/EmptyCardWithText.dart
